+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/window/GeckoWindowRequest.kt
@@ -12,8 +12,9 @@ import mozilla.components.concept.engine.window.WindowRequest
  * Gecko-based implementation of [WindowRequest].
  */
 class GeckoWindowRequest(
-    override val url: String,
-    private val engineSession: GeckoEngineSession
+    override val url: String = "",
+    private val engineSession: GeckoEngineSession,
+    override val type: WindowRequest.Type = WindowRequest.Type.OPEN
 ) : WindowRequest {
 
     override fun prepare(): EngineSession {