+++ mobile/android/android-components/components/browser/engine-system/src/main/java/mozilla/components/browser/engine/system/window/SystemWindowRequest.kt
@@ -26,7 +26,8 @@ class SystemWindowRequest(
     private val newWebView: WebView? = null,
     val openAsDialog: Boolean = false,
     val triggeredByUser: Boolean = false,
-    private val resultMsg: Message? = null
+    private val resultMsg: Message? = null,
+    override val type: WindowRequest.Type = WindowRequest.Type.OPEN
 ) : WindowRequest {
 
     override val url: String = ""