+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/SessionTest.kt
@@ -27,7 +27,6 @@ import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.engine.media.RecordingDevice
 import mozilla.components.concept.engine.permission.PermissionRequest
-import mozilla.components.concept.engine.window.WindowRequest
 import mozilla.components.support.base.observer.Consumable
 import mozilla.components.support.test.any
 import mozilla.components.support.test.argumentCaptor
@@ -684,7 +683,6 @@ class SessionTest {
         val defaultObserver = object : Session.Observer {}
         val contentPermissionRequest: PermissionRequest = mock()
         val appPermissionRequest: PermissionRequest = mock()
-        val windowRequest: WindowRequest = mock()
 
         defaultObserver.onUrlChanged(session, "")
         defaultObserver.onTitleChanged(session, "")
@@ -703,8 +701,6 @@ class SessionTest {
         defaultObserver.onThumbnailChanged(session, spy(Bitmap::class.java))
         defaultObserver.onContentPermissionRequested(session, contentPermissionRequest)
         defaultObserver.onAppPermissionRequested(session, appPermissionRequest)
-        defaultObserver.onOpenWindowRequested(session, windowRequest)
-        defaultObserver.onCloseWindowRequested(session, windowRequest)
         defaultObserver.onWebAppManifestChanged(session, mock())
         defaultObserver.onMediaAdded(session, emptyList(), mock())
         defaultObserver.onMediaRemoved(session, emptyList(), mock())
@@ -766,62 +762,6 @@ class SessionTest {
 
         assertTrue(appPermissionCallbackExecuted)
         assertTrue(session.appPermissionRequest.isConsumed())
-    }
-
-    @Test
-    fun `window requests will be set on session if no observer consumes them`() {
-        val openWindowRequest: WindowRequest = mock()
-        val closeWindowRequest: WindowRequest = mock()
-
-        val session = Session("https://www.mozilla.org")
-        session.openWindowRequest = Consumable.from(openWindowRequest)
-        session.closeWindowRequest = Consumable.from(closeWindowRequest)
-        assertFalse(session.openWindowRequest.isConsumed())
-        assertFalse(session.closeWindowRequest.isConsumed())
-
-        var createWindowRequestIsSet = false
-        var closeWindowRequestIsSet = false
-        session.openWindowRequest.consume {
-            createWindowRequestIsSet = true
-            true
-        }
-        session.closeWindowRequest.consume {
-            closeWindowRequestIsSet = true
-            true
-        }
-        assertTrue(createWindowRequestIsSet)
-        assertTrue(closeWindowRequestIsSet)
-    }
-
-    @Test
-    fun `window requests will not be set on session if consumed by observer`() {
-        var openWindowRequestExecuted = false
-        var closeWindowRequestExecuted = false
-
-        val session = Session("https://www.mozilla.org")
-        session.register(object : Session.Observer {
-            override fun onOpenWindowRequested(session: Session, windowRequest: WindowRequest): Boolean {
-                openWindowRequestExecuted = true
-                return true
-            }
-
-            override fun onCloseWindowRequested(session: Session, windowRequest: WindowRequest): Boolean {
-                closeWindowRequestExecuted = true
-                return true
-            }
-        })
-
-        val createWindowRequest: WindowRequest = mock()
-        session.openWindowRequest = Consumable.from(createWindowRequest)
-
-        val closeWindowRequest: WindowRequest = mock()
-        session.closeWindowRequest = Consumable.from(closeWindowRequest)
-
-        assertTrue(openWindowRequestExecuted)
-        assertTrue(session.openWindowRequest.isConsumed())
-
-        assertTrue(closeWindowRequestExecuted)
-        assertTrue(session.closeWindowRequest.isConsumed())
     }
 
     @Test