+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/BrowserFragment.kt
@@ -12,7 +12,7 @@ import kotlinx.android.synthetic.main.fragment_browser.view.*
 import mozilla.components.feature.awesomebar.AwesomeBarFeature
 import mozilla.components.feature.awesomebar.provider.SearchSuggestionProvider
 import mozilla.components.feature.session.ThumbnailsFeature
-import mozilla.components.feature.session.WindowFeature
+import mozilla.components.feature.tabs.WindowFeature
 import mozilla.components.feature.tabs.toolbar.TabsToolbarFeature
 import mozilla.components.feature.toolbar.ToolbarAutocompleteFeature
 import mozilla.components.support.base.feature.BackHandler
@@ -69,7 +69,7 @@ class BrowserFragment : BaseBrowserFragment(), BackHandler {
             view = layout
         )
 
-        val windowFeature = WindowFeature(components.sessionManager)
+        val windowFeature = WindowFeature(components.store, components.tabsUseCases)
         lifecycle.addObserver(windowFeature)
 
         return layout